Update build.gradle
Fixes #61 

Used https://github.com/react-native-community/async-storage/blob/master/android/build.gradle for inspiration. This is a quick fix and I'm not familiar with this project so please double check that everything is ok.

I have verified that this fixes #20 in our project and the workaround that was posted there is no longer necessary. Compilation seems to complete properly.